Trait sui_package_resolver::PackageStore

pub trait PackageStore: Send + Sync + 'static {
    // Required methods
    fn version<'life0, 'async_trait>(
        &'life0 self,
        id: AccountAddress
    ) -> Pin<Box<dyn Future<Output = Result<SequenceNumber>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn fetch<'life0, 'async_trait>(
        &'life0 self,
        id: AccountAddress
    ) -> Pin<Box<dyn Future<Output = Result<Arc<Package>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
}

Interface to abstract over access to a store of live packages. Used to override the default store during testing.

Required Methods

fn version<'life0, 'async_trait>( &'life0 self, id: AccountAddress ) -> Pin<Box<dyn Future<Output = Result<SequenceNumber>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Latest version of the object at id.

fn fetch<'life0, 'async_trait>( &'life0 self, id: AccountAddress ) -> Pin<Box<dyn Future<Output = Result<Arc<Package>>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Read package contents. Fails if id is not an object, not a package, or is malformed in some way.