sui_package_resolver
Trait PackageStore

pub trait PackageStore:
    Send
    + Sync
    + 'static {
    // Required method
    fn fetch<'life0, 'async_trait>(
        &'life0 self,
        id: AccountAddress,
    ) -> Pin<Box<dyn Future<Output = Result<Arc<Package>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
}

Interface to abstract over access to a store of live packages. Used to override the default store during testing.

Required Methods

fn fetch<'life0, 'async_trait>( &'life0 self, id: AccountAddress, ) -> Pin<Box<dyn Future<Output = Result<Arc<Package>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Read package contents. Fails if id is not an object, not a package, or is malformed in some way.